scrに統合
